Refactored functions to utils modules
